Implementing Brightest-Path-Tracing into Append Points, Fix for Background ROI Bug
For brightestPath(), we are using AStarSearch to get a lineString following the brightest path that connects a rough segment (first point appended, last point appended).
To speed up the algorithm, the image passed in is reduced to a bounding box around the rough segment.
To reduce the amount of points being added to the segment (so that there are less jagged lines), the data that is returned from the AStarSearch() is fitted with a Spline. This is done with SciPy's spline interpolation functions that returns a smoothed line given a set of points. This smooth line is then rounded to maintain the original points' integer form.